ai-detector/requirements.txt
torch
transformers
opencv-python
numpy
matplotlib
Pillow